requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2>=2.8
django-cors-headers==3.5.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.0
